Revert "🔥 remove: sentry scaffolded code"
