+++ .gitignore
@@ -52,6 +52,3 @@ _posts
 /playwright/.cache/
 credentials.json
 tsconfig.tsbuildinfo
-
-# Sentry
-.sentryclirc
