+++ package.json
@@ -4,10 +4,7 @@
   "license": "MIT",
   "scripts": {
     "dev": "next dev",
-    "build": "run-s -l -c build:next generate-sitemap delete-sourcemap",
-    "build:next": "next build && next export",
-    "generate-sitemap": "next-sitemap --config sitemap.config.js",
-    "delete-sourcemap": "find ./out -name *.map -delete",
+    "build": "next build && next-sitemap --config sitemap.config.js",
     "start": "next start",
     "lint": "run-p -l -c --aggregate-output lint:*",
     "lint:eslint": "eslint --cache .",
@@ -18,7 +15,6 @@
     "test": "run-p -l -c --aggregate-output lint:* unit e2e",
     "unit": "jest",
     "e2e": "playwright test",
-    "vrt": "playwright test test/vrt",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook"
   },
@@ -33,7 +29,6 @@
   "dependencies": {
     "@google-analytics/data": "3.0.0",
     "@headlessui/react": "1.6.6",
-    "@sentry/nextjs": "7.10.0",
     "@tailwindcss/aspect-ratio": "0.4.0",
     "@tailwindcss/line-clamp": "0.4.0",
     "@tailwindcss/typography": "^0.5.4",
@@ -102,7 +97,6 @@
     "lint-staged": "13.0.3",
     "marked": "4.0.18",
     "msw": "0.44.2",
-    "next-compose-plugins": "2.2.1",
     "next-sitemap": "3.1.15",
     "npm-run-all": "4.1.5",
     "postcss": "8.4.14",
