--- sentry.properties
@@ -1,4 +1,0 @@
-defaults.url=https://sentry.io/
-defaults.org=pandashark
-defaults.project=food-blog
-cli.executable=../../../../Library/pnpm/store/v3/tmp/dlx-16193/node_modules/.pnpm/@sentry+cli@1.74.5/node_modules/@sentry/cli/bin/sentry-cli
