+++ src/components/functions/error/ErrorBoundaryBase/index.tsx
@@ -1,18 +1,12 @@
-import { memo, useEffect } from "react";
+import { memo } from "react";
 import type { FallbackProps } from "react-error-boundary";
 import { ErrorBoundary } from "react-error-boundary";
 
 import { ErrorFallback } from "@/components/organisms/ErrorFallback";
-import { sentryLog } from "@/lib/sentry";
 
 const FallbackComponent = ({ error, resetErrorBoundary }: FallbackProps) => {
   const handleReset = () => resetErrorBoundary();
-
-  useEffect(() => {
-    if (error) {
-      sentryLog(error);
-    }
-  }, [error]);
+  console.error(error);
 
   return (
     <ErrorFallback heading="Something went wrong" message="サイト上で問題が発生しました。" onReset={handleReset} />
@@ -22,7 +16,7 @@ const FallbackComponent = ({ error, resetErrorBoundary }: FallbackProps) => {
 export const ErrorBoundaryBase: React.FC<{ children: React.ReactNode; callback?: () => void }> = memo(
   ({ children, callback }) => {
     const handleReset = () => {
-      callback?.();
+      callback && callback();
     };
 
     return (
