+++ src/components/layouts/RootLayout/index.tsx
@@ -1,7 +1,7 @@
-import { memo } from "react";
+import dynamic from "next/dynamic";
+import { memo, useMemo } from "react";
 
 import { LayoutErrorBoundary } from "@/components/functions/error";
-import Footer from "@/components/organisms/Footer";
 import Header from "@/components/organisms/Header";
 import type { TCategory, TConfig } from "@/types";
 
@@ -12,13 +12,16 @@ type RootLayoutProps = {
 };
 
 export const RootLayout: React.FC<RootLayoutProps> = memo(({ config, categories, children }) => {
+  const Footer = useMemo(() => dynamic(() => import("@/components/organisms/Footer")), []);
   const { siteTitle, organization } = config;
   return (
-    <div className="mx-auto mb-16 flex min-h-screen max-w-screen-xl flex-col px-4">
-      <Header siteTitle={siteTitle} categories={categories} />
-      <LayoutErrorBoundary>{children}</LayoutErrorBoundary>
-      <Footer siteTitle={siteTitle} organization={organization} />
-    </div>
+    <>
+      <div className="mx-auto mb-16 flex min-h-screen max-w-screen-xl flex-col px-4">
+        <Header siteTitle={siteTitle} categories={categories} />
+        <LayoutErrorBoundary>{children}</LayoutErrorBoundary>
+        <Footer siteTitle={siteTitle} organization={organization} />
+      </div>
+    </>
   );
 });
