--- src/lib/sentry/index.ts
@@ -1,1 +1,0 @@
-export { sentryLog } from "./logger";
