+++ src/pages/404/index.spec.tsx
@@ -7,9 +7,9 @@ describe("pages/404", () => {
   const mockCategoryList = Object.values(mockCategories);
 
   it("OK: 初期レンダリング", async () => {
-    render(<ErrorPage config={mockConfig} categories={mockCategoryList} />);
+    const { unmount } = render(<ErrorPage config={mockConfig} categories={mockCategoryList} />);
     const h1 = screen.getByRole("heading", { level: 1 });
     expect(h1).toHaveTextContent("404 - Not Found");
-    expect(screen.getByText("ページが見つかりませんでした")).toBeInTheDocument();
+    unmount();
   });
 });
