+++ src/pages/_error/index.spec.tsx
@@ -1,12 +1,15 @@
+import { mockCategories, mockConfig } from "@mocks/data";
 import { render, screen } from "jest/test-utils";
 
 import ErrorPage from "./index.page";
 
 describe("pages/_error", () => {
+  const mockCategoryList = Object.values(mockCategories);
+
   it("OK: 初期レンダリング", async () => {
-    render(<ErrorPage pageTitle="Unhandled Error" message="Error Message" />);
+    const { unmount } = render(<ErrorPage config={mockConfig} categories={mockCategoryList} />);
     const h1 = screen.getByRole("heading", { level: 1 });
     expect(h1).toHaveTextContent("Unhandled Error");
-    expect(screen.getByText("Error Message")).toBeInTheDocument();
+    unmount();
   });
 });
