+++ src/pages/api/preview/index.page.ts
@@ -1,4 +1,3 @@
-import { withSentry } from "@sentry/nextjs";
 import type { NextApiRequest, NextApiResponse } from "next";
 
 import { fetchArticle } from "@/api/fetchArticles";
@@ -28,4 +27,4 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   res.end("Preview mode enabled");
 };
 
-export default withSentry(handler);
+export default handler;
